+++ src/components/shareable-url-modal/loading-view/loading-view.js
@@ -1,0 +1,9 @@
+import LoadingIcon from '../../icons/loading';
+
+const LoadingView = ({ isLoading }) => (
+  <div className="shareable-url-modal__loading">
+    <LoadingIcon visible={isLoading} />
+  </div>
+);
+
+export default LoadingView;
