+++ src/components/shareable-url-modal/main-view/main-view.js
@@ -1,0 +1,187 @@
+/* eslint-disable camelcase */
+import classnames from 'classnames';
+import Dropdown from '../../ui/dropdown';
+import Button from '../../ui/button';
+import IconButton from '../../ui/icon-button';
+import InfoIcon from '../../icons/info';
+import Input from '../../ui/input';
+import MenuOption from '../../ui/menu-option';
+import Toggle from '../../ui/toggle';
+import {
+  hostingPlatforms,
+  KEDRO_VIZ_PUBLISH_AWS_DOCS_URL,
+  KEDRO_VIZ_PUBLISH_AZURE_DOCS_URL,
+  KEDRO_VIZ_PUBLISH_GCP_DOCS_URL,
+  KEDRO_VIZ_PUBLISH_DOCS_URL,
+} from '../../../config';
+
+const renderTextContent = (toggleValue, setTogleValue) => {
+  return (
+    <div className="shareable-url-modal__content-wrapper">
+      <div className="shareable-url-modal__content-title">
+        Publish and Share Kedro-Viz
+      </div>
+      <h2 className="shareable-url-modal__content-description-title">
+        Prerequisite:{' '}
+      </h2>
+      <p className="shareable-url-modal__content-description">
+        Deploying and hosting Kedro-Viz requires access keys or user
+        credentials, depending on the chosen service provider. To use this
+        feature, please add your access keys or credentials as environment
+        variables in your project. More information can be found in the{' '}
+        <a
+          target="_blank"
+          rel="noopener noreferrer"
+          href={KEDRO_VIZ_PUBLISH_DOCS_URL}
+        >
+          documentation
+        </a>
+        .
+      </p>
+      <h2 className="shareable-url-modal__content-description-title">
+        Disclaimer:{' '}
+      </h2>
+      <p className="shareable-url-modal__content-description">
+        Disclaimer Kedro-Viz contains preview data for multiple datasets. You
+        can enable or disable all previews when publishing Kedro-Viz.
+      </p>
+      <div className="shareable-url-modal__content-preview-dataset">
+        All dataset previews
+        <Toggle
+          className="shareable-url-modal__content-toggle"
+          title={toggleValue ? 'On' : 'Off'}
+          checked={toggleValue}
+          onChange={() => setTogleValue((prev) => !prev)}
+        />
+      </div>
+    </div>
+  );
+};
+
+const MainView = ({
+  handleModalClose,
+  handleSubmit,
+  inputValues,
+  isFormDirty,
+  onChange,
+  setTogleValue,
+  toggleValue,
+  visible,
+}) => {
+  const { platform, bucket_name, endpoint } = inputValues || {};
+
+  return (
+    <>
+      <div className="shareable-url-modal__content-form-wrapper">
+        {renderTextContent(toggleValue, setTogleValue)}
+        <div className="shareable-url-modal__form-wrapper">
+          <p className="shareable-url-modal__form-wrapper-title">
+            Enter the required information.
+          </p>
+          <div className="shareable-url-modal__input-wrapper">
+            <div className="shareable-url-modal__input-label">
+              Hosting platform
+            </div>
+            <Dropdown
+              defaultText={platform && hostingPlatforms[platform]}
+              placeholderText={!platform ? 'Select a hosting platform' : null}
+              onChanged={onChange}
+              width={null}
+            >
+              {Object.entries(hostingPlatforms).map(([value, label]) => (
+                <MenuOption
+                  className={classnames({
+                    'pipeline-list__option--active': platform === value,
+                  })}
+                  key={value}
+                  primaryText={label}
+                  value={value}
+                />
+              ))}
+            </Dropdown>
+          </div>
+          <div className="shareable-url-modal__input-wrapper">
+            <div className="shareable-url-modal__input-label">Bucket name</div>
+            <Input
+              defaultValue={bucket_name}
+              onChange={(value) => onChange('bucket_name', value)}
+              placeholder="Enter name"
+              resetValueTrigger={visible}
+              size="small"
+              type="input"
+              dataTest={'bucket_name'}
+            />
+          </div>
+          <div className="shareable-url-modal__input-wrapper">
+            <div className="shareable-url-modal__endpoint-url-wrapper">
+              <div className="shareable-url-modal__input-label">
+                Endpoint URL
+              </div>
+              <IconButton
+                ariaLabel="The endpoint URL information"
+                className="shareable-url-modal__information-icon"
+                labelText={
+                  <p>
+                    The endpoint URL is the link to where your Kedro-Viz will be
+                    hosted. For information on obtaining the endpoint URL,
+                    please refer to the documentation for{' '}
+                    <a
+                      className="shareable-url-modal__input-label-text"
+                      href={KEDRO_VIZ_PUBLISH_AWS_DOCS_URL}
+                      rel="noreferrer"
+                      target="_blank"
+                    >
+                      AWS
+                    </a>
+                    ,{' '}
+                    <a
+                      className="shareable-url-modal__input-label-text"
+                      href={KEDRO_VIZ_PUBLISH_AZURE_DOCS_URL}
+                      rel="noreferrer"
+                      target="_blank"
+                    >
+                      Azure
+                    </a>
+                    ,{' '}
+                    <a
+                      className="shareable-url-modal__input-label-text"
+                      href={KEDRO_VIZ_PUBLISH_GCP_DOCS_URL}
+                      rel="noreferrer"
+                      target="_blank"
+                    >
+                      GCP
+                    </a>
+                  </p>
+                }
+                icon={InfoIcon}
+              />
+            </div>
+            <Input
+              defaultValue={endpoint}
+              onChange={(value) => onChange('endpoint', value)}
+              placeholder="Enter url"
+              resetValueTrigger={visible}
+              size="small"
+              type="input"
+              dataTest={'endpoint_name'}
+            />
+          </div>
+        </div>
+      </div>
+      <div className="shareable-url-modal__button-wrapper">
+        <Button mode="secondary" onClick={handleModalClose} size="small">
+          Cancel
+        </Button>
+        <Button
+          disabled={!Object.values(isFormDirty).every((value) => value)}
+          size="small"
+          onClick={handleSubmit}
+        >
+          Publish
+        </Button>
+      </div>
+    </>
+  );
+};
+
+export default MainView;
