+++ src/components/shareable-url-modal/success-view/success-view.js
@@ -1,0 +1,21 @@
+import UrlBox from '../url-box/url-box';
+
+const SuccessView = ({
+  handleResponseUrl,
+  onClick,
+  responseUrl,
+  showCopied,
+}) => {
+  return responseUrl ? (
+    <div className="shareable-url-modal__result">
+      <UrlBox
+        url={responseUrl}
+        onClick={onClick}
+        href={handleResponseUrl()}
+        showCopiedText={showCopied}
+      />
+    </div>
+  ) : null;
+};
+
+export default SuccessView;
