+++ src/config.js
@@ -156,6 +156,18 @@ export const hostingPlatforms = {
   azure: 'Microsoft Azure',
 };
 
+export const shareableUrlMessages = (status, info = '') => {
+  const messages = {
+    failure: 'Something went wrong. Please try again later.',
+    loading: 'Shooting your files through space. Sit tight...',
+    success:
+      'The deployment has been successful and Kedro-Viz is hosted via the link below..',
+    incompatible: `Publishing Kedro-Viz is only supported with fsspec>=2023.9.0. You are currently on version ${info}.\n\nPlease upgrade fsspec to a supported version and ensure you're using Kedro 0.18.2 or above.`,
+  };
+
+  return messages[status];
+};
+
 export const inputKeyToStateKeyMap = {
   // eslint-disable-next-line camelcase
   bucket_name: 'hasBucketName',
